Security, error handling and frontEnd (#87)
* Fix uncaught in promise and internal error when uploading scripts. Fix #78

* new .env bakc in file moving some env vars from root to back env

* deleting flags and v-if -> replaced by addAlertMessage in editView and registerView. Update some errors in back

* Front omprovments for editView. Inprogress

* Error handling, checks server-side and front end for editView

* front in homeview

* Homepage almost beautiful

* insert meteorite on homepage and improve flex display

* front for 2fa

* input styles in an extern scss file

* added favicon, move register view into a component, and stop redirecting user during registration -> all stuff is done on homeview

* Edit view fornt end improved. Refactoring css elements

* 2fa svg improvments

* rename .scss files and add border to switch

* Front improvments for profile view. Add few svgs and reorganize some css

* Profile almost done. User cards done.

* add logout in navbar and editIcon on profile

* starting front on game page

* add cool gif to explore for lobby illustration

* Game page fornt end. New disposition and new assets. Need to style custom games

* finish game page front and add pizz.jpeg

* adjust container height for editView

* add new buttons to profile page

* Deleting lobby animation during game

* Delete logo.svg

* Delete twoFaEnabled.svg

+++ back/app/src/authentication/authentication.service.ts
@@ -28,14 +28,14 @@ export class AuthenticationService {
     } catch (error) {
       await queryRunner.rollbackTransaction();
       console.log(error);
-      if (imageDto) {
+      if (imageDto && imageDto.filename != 'pizz.jpeg') {
         unlink(imageDto.path, (err) => {
           if (err) throw err;
           console.log(imageDto.path + ' has been deleted');
         });
       }
       if (error?.code === PostgresErrorCode.UniqueViolation) {
-        throw new UserAlreadyExistException();
+        throw new UnauthorizedException('Nickname already exists');
       }
       throw new InternalServerErrorException();
     } finally {
